refactor(platform): move set-loop and fibonacci delays into utils
Move setLoop() and DEFAULT_FIBONACCI_DELAYS_MS from polling.ts into
utils.ts alongside the other centralised try/catch utilities.
polling.ts re-exports setLoop for backward compatibility.
