+++ turbo/apps/platform/src/signals/utils.ts
@@ -1,5 +1,6 @@
 import { command, state, type Command } from "ccstate";
 import type { CSSProperties } from "react";
+import { delay } from "signal-timers";
 import { IN_VITEST } from "../env.ts";
 import { logger } from "./log.ts";
 
@@ -157,6 +158,49 @@ export async function bestEffort(p: Promise<unknown>): Promise<void> {
     await p;
   } catch (error) {
     throwIfAbort(error);
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Polling loop with fibonacci backoff
+// ---------------------------------------------------------------------------
+
+export const DEFAULT_FIBONACCI_DELAYS_MS = [
+  1000, 1000, 2000, 3000, 5000, 8000, 13_000, 21_000, 34_000, 55_000, 60_000,
+] as const;
+
+/**
+ * Run `loopBody` in a loop with `interval` between iterations.
+ * Transient (non-abort) errors trigger fibonacci backoff retries.
+ * Resolves when `loopBody` returns `true` (done) or rejects on abort.
+ */
+export async function setLoop(
+  loopBody: (signal: AbortSignal) => Promise<boolean> | boolean,
+  interval: number,
+  signal: AbortSignal,
+  fibDelays: readonly number[] = DEFAULT_FIBONACCI_DELAYS_MS,
+): Promise<void> {
+  let fibIndex = 0;
+  while (true) {
+    // eslint-disable-next-line no-restricted-syntax -- polling loop requires try/catch for transient error retry with backoff
+    try {
+      const done = await loopBody(signal);
+      if (done) {
+        return;
+      }
+      fibIndex = 0;
+      await delay(interval, { signal });
+    } catch (error) {
+      throwIfAbort(error);
+      const backoff =
+        fibDelays[Math.min(fibIndex, fibDelays.length - 1)] ?? 60_000;
+      L.warn(
+        `setLoop: transient error (attempt ${fibIndex + 1}), retrying in ${backoff}ms`,
+        error,
+      );
+      fibIndex++;
+      await delay(backoff, { signal });
+    }
   }
 }
